+++ resources/views/admin/country/form.blade.php
@@ -12,6 +12,11 @@
                                 {{ session('status') }}
                             </div>
                         @endif
+                        @if(Session::has('success'))
+                            <div class="alert alert-success">
+                                {{Session::get('success')}}
+                            </div>
+                        @endif
                         <!-- kiểm tra nếu không tồn tại dùng form store -->
                         @if(!isset($country))
                             {!! Form::open(['route'=>'country.store','method'=>'POST']) !!}
@@ -21,24 +26,35 @@
                                 <div class="form-group text-white">
                                     {!! Form::label('title', 'Tên Quốc Gia', []) !!}
                                     {!! Form::text('title', isset($country) ? $country->title : '', ['class' => 'form-control', 'placeholder' =>'Nhập nội dung','id' => 'slug','onkeyup' => 'ChangeToSlug()']) !!}
+                                    @error('title')
+                                            <div class="alert alert-danger">{{ $message }}</div>
+                                    @enderror
                                 </div>
                                 <div class="form-group text-white">
                                     {!! Form::label('slug', 'Slug', []) !!}
-                                    {!! Form::text('slug', isset($country) ? $country->slug : '', ['class' => 'form-control', 'placeholder' =>'Nhập nội dung','id' => 'convert_slug']) !!}
+                                    {!! Form::text('slug', isset($country) ? $country->slug : '', ['class' => 'form-control', 'placeholder' =>'auto','id' => 'convert_slug']) !!}
                                 </div>
                                 <div class="form-group text-white">
                                     {!! Form::label('description','Mô tả', []) !!}
                                     {!! Form::textarea('description', isset($country) ? $country->description : '', ['style' => 'resize:none','class' => 'form-control', 'placeholder' =>'Nhập nội dung','id' => 'content']) !!}
+                                    @error('description')
+                                            <div class="alert alert-danger">{{ $message }}</div>
+                                    @enderror
                                 </div>
                                 <div class="form-group text-white">
                                     {!! Form::label('Active','Trạng thái', []) !!}
                                     {!! Form::select('status', ['1' => 'Hiển thị danh mục phim' , '0' => 'Không hiện'], isset($country) ? $country->status : '', ['class' => 'form-control', 'placeholder' =>'Nhập nội dung']) !!}
+                                    @error('status')
+                                            <div class="alert alert-danger">{{ $message }}</div>
+                                    @enderror
                                 </div>
-                                @if(!isset($country))
-                                        {!! Form::submit('Thêm dữ liệu', ['class' => 'btn btn-primary']) !!}
-                                @else
-                                
-                                        {!! Form::submit('Cập nhật dữ liệu', ['class' => 'btn btn-primary']) !!}
+                                @if(Auth::user()->role == 'admin' || Auth::user()->role == 'superadmin' || Auth::user()->role == 'manage' && Auth::user()->status == '1')
+                                    @if(!isset($country))
+                                            {!! Form::submit('Thêm dữ liệu', ['class' => 'btn btn-primary']) !!}
+                                    @else
+                                    
+                                            {!! Form::submit('Cập nhật dữ liệu', ['class' => 'btn btn-primary']) !!}
+                                    @endif
                                 @endif
                             {!! Form::close() !!}
                     </div>
@@ -66,14 +82,16 @@
                                                     Không hiển thị
                                                 @endif
                                             </td>
-                                            <td>
-                                                {!! Form::open(['method'=>'delete','route' => ['country.destroy', $gen->id], 'onsubmit' => 'return confirm("Bạn có muốn xóa?")']) !!}
-                                                    {{ Form::button('<i class="fa fa-trash"></i>', ['type' => 'submit', 'class' => 'btn btn-dark btn-sm', 'style' => 'height:40px; width:40px'] )  }}
-                                                {!! Form::close() !!}                                
-                                            </td>
-                                            <td>
-                                                <a href="{{route('country.edit', $gen->id)}}" class="btn btn-warning" style = "height:40px; width:40px"><i class="fa-solid fa-pen"></i></a>
-                                            </td>
+                                            @if(Auth::user()->role == 'admin' || Auth::user()->role == 'superadmin' || Auth::user()->role == 'manage' && Auth::user()->status == '1')
+                                                <td>
+                                                    {!! Form::open(['method'=>'delete','route' => ['country.destroy', $gen->id], 'onsubmit' => 'return confirm("Bạn có muốn xóa?")']) !!}
+                                                        {{ Form::button('<i class="fa fa-trash"></i>', ['type' => 'submit', 'class' => 'btn btn-dark btn-sm', 'style' => 'height:40px; width:40px'] )  }}
+                                                    {!! Form::close() !!}                                
+                                                </td>
+                                                <td>
+                                                    <a href="{{route('country.edit', $gen->id)}}" class="btn btn-warning" style = "height:40px; width:40px"><i class="fa-solid fa-pen"></i></a>
+                                                </td>
+                                            @endif
                                     </tr>
                                 @endforeach
                                 
@@ -82,7 +100,8 @@
                     <form action="{{url('/downloadPDF')}}" method="get">
                             @csrf
                             <button class="btn btn-primary">Tải về</button>
-                    </form>     
+                    </form>    
+                    
                     <div>
                         {{$lstCountry->links("pagination::bootstrap-5")}}
                     </div>
